+++ libfive/test/CMakeLists.txt
@@ -15,6 +15,7 @@ set(SRCS main.cpp
     eval_point.cpp
     feature.cpp
     heightmap.cpp
+    indexes.cpp
     marching.cpp
     mesh.cpp
     neighbors.cpp
